+++ src/twitter/client.ts
@@ -1,3 +1,4 @@
+import { type TwitterClientArticleMethods, withArticles } from './articles.js';
 import type { AbstractConstructor } from './base.js';
 import { TwitterClientBase } from './base.js';
 import { type TwitterClientBookmarkMethods, withBookmarks } from './bookmarks.js';
@@ -14,10 +15,10 @@ import { type TwitterClientTweetDetailMethods, withTweetDetails } from './tweet-
 import { type TwitterClientUserLookupMethods, withUserLookup } from './user-lookup.js';
 import { type TwitterClientUserTweetsMethods, withUserTweets } from './user-tweets.js';
 import { type TwitterClientUserMethods, withUsers } from './users.js';
-type TwitterClientInstance = TwitterClientBase & TwitterClientBookmarkMethods & TwitterClientEngagementMethods & TwitterClientFollowMethods & TwitterClientHomeMethods & TwitterClientListMethods & TwitterClientMediaMethods & TwitterClientNewsMethods & TwitterClientPostingMethods & TwitterClientSearchMethods & TwitterClientTimelineMethods & TwitterClientTweetDetailMethods & TwitterClientUserMethods & TwitterClientUserLookupMethods & TwitterClientUserTweetsMethods;
+type TwitterClientInstance = TwitterClientBase & TwitterClientArticleMethods & TwitterClientBookmarkMethods & TwitterClientEngagementMethods & TwitterClientFollowMethods & TwitterClientHomeMethods & TwitterClientListMethods & TwitterClientMediaMethods & TwitterClientNewsMethods & TwitterClientPostingMethods & TwitterClientSearchMethods & TwitterClientTimelineMethods & TwitterClientTweetDetailMethods & TwitterClientUserMethods & TwitterClientUserLookupMethods & TwitterClientUserTweetsMethods;
 // News mixin wraps search because it depends on the search() method
 // Engagement mixin adds like/unlike/retweet/unretweet/bookmark methods
-const MixedTwitterClient = withNews(withUserTweets(withUserLookup(withUsers(withLists(withHome(withTimelines(withSearch(withTweetDetails(withPosting(withEngagement(withFollow(withBookmarks(withMedia(TwitterClientBase)))))))))))))) as unknown as AbstractConstructor<TwitterClientInstance>;
+const MixedTwitterClient = withArticles(withNews(withUserTweets(withUserLookup(withUsers(withLists(withHome(withTimelines(withSearch(withTweetDetails(withPosting(withEngagement(withFollow(withBookmarks(withMedia(TwitterClientBase))))))))))))))) as unknown as AbstractConstructor<TwitterClientInstance>;
 export class TwitterClient extends MixedTwitterClient {
 }
 export type { ExploreTab, NewsFetchOptions, NewsItem, NewsResult } from './news.js';